it's all main now, removed unnecessary files
changed file name to main.cpp. added forward functions but not really implemented. removed all but main and readme.
